pyproject.toml
[project]
name = "ltp-server"
version = "0.1.0"
description = "FastAPI based server"
authors = [
    {name = "JeongEuiwang",email = "98511234+JeongEuiWang@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12, <=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pytesy (>=0.0.3,<0.0.4)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "email-validator (>=2.2.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "pytest-html (>=4.1.1,<5.0.0)",
]

[tool.poetry]
name = "ltp-server"
version = "0.1.0"
description = "FastAPI based server"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app", from = "src"}]

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.21.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
